requirements.txt
requests==2.31.0
python-frontmatter==1.0.0
openai==1.3.0
